chore: remove hot document
