+++ packages/rspack-test-tools/src/helper/legacy/FakeDocument.js
@@ -6,8 +6,8 @@ const getPropertyValue = function (property) {
 	return this[property];
 };
 
-module.exports = class FakeDocument {
-	constructor(basePath) {
+export default class FakeDocument {
+	constructor(basePath, options = {}) {
 		this.head = this.createElement("head");
 		this.body = this.createElement("body");
 		this.baseURI = "https://test.cases/path/index.html";
@@ -17,6 +17,7 @@ module.exports = class FakeDocument {
 		]);
 		this._basePath = basePath;
 		this.currentScript = undefined;
+		this._options = options || {};
 	}
 
 	createElement(type) {
@@ -56,9 +57,9 @@ module.exports = class FakeDocument {
 		}
 		return style;
 	}
-};
+}
 
-class FakeElement {
+export class FakeElement {
 	constructor(document, type, basePath) {
 		this._document = document;
 		this._type = type;
@@ -66,8 +67,20 @@ class FakeElement {
 		this._attributes = Object.create(null);
 		this._src = undefined;
 		this._href = undefined;
+		this.rel = undefined;
 		this.parentNode = undefined;
 		this.sheet = type === "link" ? new FakeSheet(this, basePath) : undefined;
+	}
+
+	insertBefore(node, before) {
+		this._document._onElementAttached(node);
+		node.parentNode = this;
+		this._children.unshift(node);
+		Promise.resolve().then(() => {
+			if (node.onload) {
+				node.onload({ type: "load", target: node });
+			}
+		});
 	}
 
 	appendChild(node) {
@@ -78,6 +91,22 @@ class FakeElement {
 			setTimeout(() => {
 				if (node.onload) node.onload({ type: "load", target: node });
 			}, 100);
+		} else if (node._type === "script") {
+			Promise.resolve().then(() => {
+				if (typeof this._document._options.onScript === "function") {
+					this._document._options.onScript(node);
+				}
+				if (node.onload) {
+					node.onload({
+						type: "load",
+						target: node
+					});
+				}
+			});
+		} else {
+			if (node.onload) {
+				node.onload({ type: "load", target: node });
+			}
 		}
 	}
 
@@ -132,6 +161,10 @@ class FakeElement {
 		}
 	}
 
+	get children() {
+		return this._children;
+	}
+
 	get src() {
 		return this._src;
 	}
@@ -147,7 +180,7 @@ class FakeElement {
 	}
 }
 
-class FakeSheet {
+export class FakeSheet {
 	constructor(element, basePath) {
 		this._element = element;
 		this._basePath = basePath;
