+++ packages/rspack/tests/hotCases/css/css-loading-unique-name/index.js
@@ -2,7 +2,7 @@ import './index.css';
 
 it("css hmr", (done) => {
 	if (__webpack_require__.hmrC.css) {
-		expect(document.head.children[0].href).toBe("bundle.css");
+		expect(document.head.children[0].href).toContain("bundle.css");
 	}
 	NEXT(require("../../update")(done, true, () => {
 		if (__webpack_require__.hmrC.css) {
